+++ FR/index.html
@@ -194,7 +194,6 @@
 	<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha512-5WvZa4N7Jq3TVNCp4rjcBMlc6pT3lZ7gVxjtI6IkKW+uItSa+rFgtFljvZnCxQGj8SUX5DHraKE6Mn/4smK1Cg==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
 	<script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap/4.5.3/js/bootstrap.min.js" integrity="sha512-8qmis31OQi6hIRgvkht0s6mCOittjMa9GMqtK9hes5iEQBQE/Ca6yGE5FsW36vyipGoWQswBj/QBm2JR086Rkw==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
 	<script src="https://cdnjs.cloudflare.com/ajax/libs/PapaParse/5.0.2/papaparse.min.js" integrity="sha512-z2MPSoqfIIvG4xSBF385/Vlsj3z4k1KbRkAFi2Mb5JE3LhUbY+rgX6PLnKJIsdnuYplZf7HVpdy8yg0YCbC+8w==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
-	<script type="text/javascript" src="../recipe.js"></script>
-	<script type="text/javascript" src ="../form.js"></script>
+	<script type="module" src="../app.js"></script>
 	<script type="text/javascript" src="../libs/scrolling-nav/js/scrolling-nav.js"></script>
 </html>
